demo_fastapi/requirements.txt
anyio==3.6.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.89.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.32
packaging
pathspec==0.11.0
platformdirs==2.6.2
pluggy==1.0.0
protobuf==3.20.3
pydantic==1.10.4
pytest==7.2.1
python-dotenv==0.21.1
requests==2.28.2
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
